+++ tests/types/test_temporal_entity.py
@@ -2,7 +2,6 @@ from datetime import date, datetime
 from typing import Any
 
 import pytest
-from pandas._libs.tslibs.parsing import DateParseError
 from pytz import timezone
 
 from mex.common.types import (
@@ -28,11 +27,12 @@ from mex.common.types import (
         (
             (1.3, "foo", 42),
             {},
-            "Temporal entity takes a single str, date, datetime or TemporalEntity argument or up to 7 integers",
+            "Temporal entity takes a single str, date, datetime or TemporalEntity "
+            "argument or up to 7 integers",
         ),
     ],
 )
-def test_timestamp_parsing_errors(
+def test_temporal_entity_type_errors(
     args: tuple[Any], kwargs: dict[str, Any], message: str
 ) -> None:
     with pytest.raises(TypeError, match=message):
@@ -40,63 +40,166 @@ def test_timestamp_parsing_errors(
 
 
 @pytest.mark.parametrize(
+    ("cls", "args", "kwargs", "error"),
+    [
+        (
+            YearMonthDayTime,
+            ("2001-04-24",),
+            {},
+            "Expected precision level to be one of 'hour', 'minute', 'second', 'microsecond'",
+        ),
+        (
+            YearMonthDay,
+            ("1999-02",),
+            {},
+            "Expected precision level to be 'day'",
+        ),
+        (
+            YearMonth,
+            ("1999-02",),
+            {"precision": TemporalEntityPrecision.DAY},
+            "Expected precision level to be one of 'year', 'month'",
+        ),
+        (
+            YearMonthDay,
+            ("2022-02-31",),
+            {},
+            "day is out of range for month",
+        ),
+        (
+            YearMonthDay,
+            ("2023-02-29",),
+            {},
+            "day is out of range for month",
+        ),
+        (
+            YearMonth,
+            ("1987-13",),
+            {},
+            "month must be in 1..12",
+        ),
+        (
+            YearMonthDayTime,
+            ("2018-03-02T12:61:01Z",),
+            {},
+            "minute must be in 0..59",
+        ),
+    ],
+    ids=[
+        "too precise",
+        "not precise enough",
+        "wrong explicit precision",
+        "not a valid day in any year",
+        "not a valid leap day",
+        "not a valid month",
+        "not a valid minute",
+    ],
+)
+def test_temporal_entity_value_errors(
+    cls: type[TemporalEntity], args: tuple[Any], kwargs: dict[str, Any], error: str
+) -> None:
+    with pytest.raises(ValueError, match=error):
+        cls(*args, **kwargs)
+
+
+@pytest.mark.parametrize(
     ("value", "message"),
     [(object(), "Cannot parse <class 'object'> as TemporalEntity")],
 )
-def test_timestamp_validation_errors(value: Any, message: str) -> None:
+def test_temporal_entity_validation_errors(value: Any, message: str) -> None:
     with pytest.raises(TypeError, match=message):
         TemporalEntity.validate(value)
 
 
 @pytest.mark.parametrize(
-    ("args", "kwargs", "expected"),
+    ("cls", "args", "kwargs", "expected"),
     [
-        ((), {}, "1970"),
-        ((2014,), {}, "2014"),
-        ((2009, 12), {}, "2009-12"),
-        ((1994, 12, 30), {}, "1994-12-30"),
-        ((1999, 1, 20, 22, 58, 17), {}, "1999-01-20T21:58:17Z"),
-        ((1999, 1, 20, 22, 58, 17), {"tzinfo": CET}, "1999-01-20T21:58:17Z"),
-        ((1999, 1, 20, 22, 58, 17), {"tzinfo": UTC}, "1999-01-20T22:58:17Z"),
-        (("1994-12-30",), {}, "1994-12-30"),
+        (TemporalEntity, (), {}, 'TemporalEntity("1970")'),
+        (TemporalEntity, (2014,), {}, 'TemporalEntity("2014")'),
+        (TemporalEntity, (2009, 12), {}, 'TemporalEntity("2009-12")'),
+        (TemporalEntity, (1994, 12, 30), {}, 'TemporalEntity("1994-12-30")'),
         (
+            TemporalEntity,
+            (1999, 1, 20, 22, 58, 17),
+            {},
+            'TemporalEntity("1999-01-20T21:58:17Z")',
+        ),
+        (
+            TemporalEntity,
+            (1999, 1, 20, 22, 58, 17),
+            {"tzinfo": CET},
+            'TemporalEntity("1999-01-20T21:58:17Z")',
+        ),
+        (
+            TemporalEntity,
+            (1999, 1, 20, 22, 58, 17),
+            {"tzinfo": UTC},
+            'TemporalEntity("1999-01-20T22:58:17Z")',
+        ),
+        (TemporalEntity, ("1994-12-30",), {}, 'TemporalEntity("1994-12-30")'),
+        (
+            TemporalEntity,
             ("1999-01-20T22:58:17Z",),
             {},
-            "1999-01-20T22:58:17Z",
+            'TemporalEntity("1999-01-20T22:58:17Z")',
         ),
         (
+            TemporalEntity,
             ("1999-01-20T22",),
             {},
-            "1999-01-20T21:00:00Z",
+            'TemporalEntity("1999-01-20T21:00:00Z")',
         ),
         (
+            TemporalEntity,
             ("2016-06-10T21:42:24.76073899Z",),
             {},
-            "2016-06-10T21:42:24Z",
+            'TemporalEntity("2016-06-10T21:42:24Z")',
         ),
         (
+            TemporalEntity,
             (date(2020, 3, 22),),
             {},
-            "2020-03-22",
+            'TemporalEntity("2020-03-22")',
         ),
         (
+            TemporalEntity,
             (datetime(2020, 3, 22, 14, 30, 58),),
             {},
-            "2020-03-22T13:30:58Z",
+            'TemporalEntity("2020-03-22T13:30:58Z")',
         ),
         (
+            TemporalEntity,
             (
                 datetime(
                     2020, 3, 22, 14, 30, 58, tzinfo=timezone("America/Los_Angeles")
                 ),
             ),
             {},
-            "2020-03-22T22:23:58Z",
+            'TemporalEntity("2020-03-22T22:23:58Z")',
         ),
         (
+            TemporalEntity,
             (TemporalEntity(2004, 11),),
             {},
-            "2004-11",
+            'TemporalEntity("2004-11")',
+        ),
+        (
+            YearMonthDayTime,
+            (YearMonthDayTime(2004, 11, 21, 19, 59, tzinfo=timezone("UTC")),),
+            {},
+            'YearMonthDayTime("2004-11-21T19:59:00Z")',
+        ),
+        (
+            TemporalEntity,
+            (datetime(2004, 11, 19, 00, 00),),
+            {"precision": TemporalEntityPrecision.DAY},
+            'TemporalEntity("2004-11-19")',
+        ),
+        (
+            YearMonth,
+            (datetime(2004, 11, 19, 00, 00),),
+            {"precision": TemporalEntityPrecision.YEAR},
+            'YearMonth("2004")',
         ),
     ],
     ids=[
@@ -114,17 +217,20 @@ def test_timestamp_validation_errors(value: Any, message: str) -> None:
         "date",
         "datetime",
         "pacific time",
-        "timestamp",
+        "temporal entity",
+        "sub class",
+        "temporal entity with precision",
+        "sub class with precision",
     ],
 )
-def test_timestamp_parsing(
-    args: tuple[Any], kwargs: dict[str, Any], expected: str
+def test_temporal_entity_parsing(
+    cls: type[TemporalEntity], args: tuple[Any], kwargs: dict[str, Any], expected: str
 ) -> None:
-    timestamp = TemporalEntity(*args, **kwargs)
-    assert str(timestamp) == expected
+    temporal_entity = cls(*args, **kwargs)
+    assert repr(temporal_entity) == expected
 
 
-def test_timestamp_eq() -> None:
+def test_temporal_entity_eq() -> None:
     assert TemporalEntity(2004) == TemporalEntity("2004")
     assert TemporalEntity(2004, 11) == TemporalEntity(2004, 11)
     assert TemporalEntity(2004, 11, 2) == "2004-11-02"
@@ -134,7 +240,7 @@ def test_timestamp_eq() -> None:
     assert TemporalEntity(2005) != object()
 
 
-def test_timestamp_gt() -> None:
+def test_temporal_entity_gt() -> None:
     assert TemporalEntity(2004) > TemporalEntity("2003")
     assert TemporalEntity(2004, 11) < "2013-10-02"
     assert TemporalEntity(2004, 11) <= TemporalEntity(2004, 12)
@@ -144,63 +250,21 @@ def test_timestamp_gt() -> None:
         assert TemporalEntity(2005) > object()
 
 
-def test_timestamp_str() -> None:
+def test_temporal_entity_str() -> None:
     assert str(TemporalEntity(2004, 11, 26)) == "2004-11-26"
     assert str(YearMonth(2004, 11)) == "2004-11"
     assert str(YearMonthDay(2004, 11, 26)) == "2004-11-26"
     assert str(YearMonthDayTime(2018, 3, 2, 13, 0, 1)) == "2018-03-02T12:00:01Z"
 
 
-def test_timestamp_repr() -> None:
+def test_temporal_entity_repr() -> None:
     assert (
         repr(TemporalEntity(2018, 3, 2, 13, 0, 1))
         == 'TemporalEntity("2018-03-02T12:00:01Z")'
     )
-
     assert repr(YearMonth("2022")) == 'YearMonth("2022")'
-
     assert repr(YearMonthDay("2022-10-03")) == 'YearMonthDay("2022-10-03")'
-
     assert (
         repr(YearMonthDayTime("2018-03-02T12:00:01Z"))
         == 'YearMonthDayTime("2018-03-02T12:00:01Z")'
     )
-
-
-def test_invalid_temporal_resolution_throws_error() -> None:
-    with pytest.raises(
-        ValueError,
-        match="Expected precision level to be one of "
-        "'hour', 'minute', 'second', 'microsecond'",
-    ):
-        YearMonthDayTime("2001-04-24")
-
-    with pytest.raises(ValueError, match="Expected precision level to be 'day'"):
-        YearMonthDay("1999-02")
-
-    with pytest.raises(
-        ValueError, match="Expected precision level to be one of " "'year', 'month'"
-    ):
-        YearMonth("1999-02-02")
-
-
-def test_invalid_dates_throw_error() -> None:
-    with pytest.raises(DateParseError):
-        YearMonthDay("2022-02-31")
-
-    with pytest.raises(DateParseError):
-        YearMonthDay("2023-02-29")  # not a leap year
-
-    with pytest.raises(DateParseError):
-        YearMonth("1987-13")
-
-    with pytest.raises(DateParseError):
-        YearMonthDayTime("2018-03-02T12:61:01Z")
-
-
-def test_apply_precision() -> None:
-    temporal_entity = TemporalEntity(date(2022, 12, 25))
-    temporal_entity.apply_precision(TemporalEntityPrecision.DAY)
-
-    assert temporal_entity.date_time == datetime(2022, 12, 25, 0, 0)
-    assert temporal_entity.precision == TemporalEntityPrecision.DAY